requirements.txt
beautifulsoup4==4.13.3
captum==0.7.0
elevenlabs==1.54.0
librosa==0.11.0
matplotlib==3.10.1
numpy==2.2.4
pandas==2.2.3
Pillow==11.1.0
pydub==0.25.1
python-dotenv==1.1.0
Requests==2.32.3
s3prl==0.4.17
scikit_learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
torch==2.6.0
tqdm==4.67.1
yt_dlp==2025.3.26
textgrid
